project-root/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
python-multipart==0.0.6
pydantic==2.4.2